Moein

+++ Includes/LibrariesAndDefines.hpp
@@ -8,6 +8,6 @@
 
 const uint64_t DEFALUT_P = 799333555511111LL;
 const uint64_t DEFAULT_MASK = -1;
-const uint8_t DEFAULT_N = 2;
+const uint32_t DEFAULT_N = 2;
 
 #endif
