+++ Src/SeedCreators/RandStrobeCreators/RandStrobeCreatorXorVar.cpp
@@ -1,7 +1,7 @@
 #include "RandStrobeCreatorXorVar.hpp"
 
 RandStrobeCreatorXorVar::RandStrobeCreatorXorVar(Hasher* hasher, Comparator* comparator, size_t kmer_len,
-		size_t w_min, size_t w_max, uint8_t n, uint64_t mask)
+		size_t w_min, size_t w_max, uint32_t n, uint64_t mask)
 : RandStrobeCreator(hasher, comparator, kmer_len, w_min, w_max, n, mask)
 {
 }
